Amend a few more endpoints according to lack of crated_on field
