+++ datahub/dataset/team/test/test_views.py
@@ -1,9 +1,5 @@
-from datetime import datetime
-
 import pytest
 
-from django.utils.timezone import utc
-from freezegun import freeze_time
 from rest_framework import status
 
 from rest_framework.reverse import reverse
@@ -54,22 +50,3 @@ class TestTeamDatasetViewSet(BaseDatasetViewTest):
         results = response.json()['results']
 
         assert results == sorted(results, key=lambda t: t['id'])
-
-    def test_with_updated_since_filter(self, data_flow_api_client):
-        with freeze_time('2021-01-01 12:30:00'):
-            TeamFactory()
-        with freeze_time('2022-01-01 12:30:00'):
-            team_after = TeamFactory()
-        # Define the `updated_since` date
-        updated_since_date = datetime(2021, 2, 1, tzinfo=utc).strftime('%Y-%m-%d')
-
-        # Make the request with the `updated_since` parameter
-        response = data_flow_api_client.get(self.view_url, {'updated_since': updated_since_date})
-
-        assert response.status_code == status.HTTP_200_OK
-
-        # Check that only contact created after the `updated_since` date are returned
-        expected_ids = [str(team_after.id)]
-        response_ids = [team['id'] for team in response.json()['results']]
-
-        assert response_ids == expected_ids
