+++ datahub/dataset/team/views.py
@@ -1,10 +1,9 @@
-from datahub.dataset.core.views import BaseFilterDatasetView
+from datahub.dataset.core.views import BaseDatasetView
 from datahub.dataset.team.pagination import TeamsDatasetViewCursorPagination
-from datahub.dbmaintenance.utils import parse_date
 from datahub.metadata.models import Team
 
 
-class TeamsDatasetView(BaseFilterDatasetView):
+class TeamsDatasetView(BaseDatasetView):
     """
     An APIView that provides 'get' action which queries and returns desired fields for
     Teams Dataset to be consumed by Data-flow periodically. Data-flow uses response result
@@ -13,9 +12,9 @@ class TeamsDatasetView(BaseFilterDatasetView):
 
     pagination_class = TeamsDatasetViewCursorPagination
 
-    def get_dataset(self, request):
+    def get_dataset(self):
         """Returns list of Teams Dataset records"""
-        queryset = Team.objects.values(
+        return Team.objects.values(
             'country__name',
             'disabled_on',
             'id',
@@ -23,11 +22,3 @@ class TeamsDatasetView(BaseFilterDatasetView):
             'role__name',
             'uk_region__name',
         )
-        updated_since = request.GET.get('updated_since')
-
-        if updated_since:
-            updated_since_date = parse_date(updated_since)
-            if updated_since_date:
-                queryset = queryset.filter(modified_on__gt=updated_since_date)
-
-        return queryset
